requirements.txt
Flask==2.3.2
pyodbc==4.0.39
gunicorn==21.2.0
wheel
setuptools
azure-identity
pywinrm